feat: adicionando testes restantes

+++ src/controllers/userController/delete.js
@@ -4,9 +4,16 @@ const SUCCESS_MESSAGE = "Usuário deletado";
 /**
  * Deleta o usuário com o id passado.
  */
+
+const EMPTY_FIELDS_MESSAGE = "Há campos vazios!";
+
 const destroy = async (req, res) => {
 
     const userId = req.id
+
+    if (!userId) {
+        return res.status(400).json({ "error": EMPTY_FIELDS_MESSAGE })
+    }
 
     try {
         await deleteUserService(userId);
